+++ python/metatensor-learn/tests/module_map.py
@@ -12,6 +12,15 @@ from metatensor.learn.nn import ModuleMap  # noqa: E402
 from .utils import TORCH_KWARGS, single_block_tensor_torch  # noqa F411 E402
 
 
+try:
+    if torch.cuda.is_available():
+        HAS_CUDA = True
+    else:
+        HAS_CUDA = False
+except ImportError:
+    HAS_CUDA = False
+
+
 class MockModule(Module):
     def __init__(self, in_features, out_features):
         super().__init__()
@@ -23,13 +32,17 @@ class MockModule(Module):
         return self._last_layer(self._activation(self._linear(input)))
 
 
-@pytest.fixture(scope="module", autouse=True)
+@pytest.fixture(scope="function", autouse=True)
 def set_random_generator():
     """Set the random generator to same seed before each test is run.
     Otherwise test behaviour is dependend on the order of the tests
     in this file and the number of parameters of the test.
     """
     torch.random.manual_seed(122578741812)
+
+
+@pytest.fixture(scope="function", autouse=True)
+def set_default_torch_resources():
     torch.set_default_device(TORCH_KWARGS["device"])
     torch.set_default_dtype(TORCH_KWARGS["dtype"])
 
@@ -83,3 +96,84 @@ def test_module_map_single_block_tensor(
                 )
             else:
                 assert out_block.gradient(parameter).properties == out_properties[0]
+
+
+@pytest.mark.parametrize(
+    "out_properties", [None, [Labels(["a", "b"], np.array([[1, 1]]))]]
+)
+def test_module_map_to(single_block_tensor_torch, out_properties):  # noqa F811
+    """
+    Checks the `to` function of module map
+    """
+    modules = []
+    for key in single_block_tensor_torch.keys:
+        modules.append(
+            MockModule(
+                in_features=len(single_block_tensor_torch.block(key).properties),
+                out_features=5,
+            )
+        )
+    tensor_module = ModuleMap(
+        single_block_tensor_torch.keys, modules, out_properties=out_properties
+    )
+    tensor_module.to("cpu")
+    single_block_tensor_torch = single_block_tensor_torch.to(device="cpu")
+    with torch.no_grad():
+        out_tensor = tensor_module(single_block_tensor_torch)
+    assert out_tensor.device.type == "cpu"
+
+
+@pytest.mark.parametrize(
+    "out_properties", [None, [Labels(["a", "b"], np.array([[1, 1]]))]]
+)
+@pytest.mark.skipif(not HAS_CUDA, reason="requires cuda")
+def test_cuda_module_map_to(single_block_tensor_torch, out_properties):  # noqa F811
+    """
+    We set the correct default device for initialization and check if the module this
+    works once the default device has been changed. This catches cases where the default
+    device or a hard coded device is used for tensor created within the forward
+    function.
+    """
+    # cuda -> cpu
+    torch.set_default_device("cuda")
+    modules = []
+    for key in single_block_tensor_torch.keys:
+        modules.append(
+            MockModule(
+                in_features=len(single_block_tensor_torch.block(key).properties),
+                out_features=5,
+            )
+        )
+
+    # labels are always on cpu in metatensor-core so we don't need to move them to
+    # default device
+    tensor_module = ModuleMap(
+        single_block_tensor_torch.keys, modules, out_properties=out_properties
+    )
+    tensor_module.to("cpu")
+    single_block_tensor_torch = single_block_tensor_torch.to(device="cpu")
+    with torch.no_grad():
+        out_tensor = tensor_module(single_block_tensor_torch)
+    assert out_tensor.device.type == "cpu"
+
+    # cpu -> cuda
+    torch.set_default_device("cpu")
+    modules = []
+    for key in single_block_tensor_torch.keys:
+        modules.append(
+            MockModule(
+                in_features=len(single_block_tensor_torch.block(key).properties),
+                out_features=5,
+            )
+        )
+    # labels are always on cpu in metatensor-core so we don't need to move them to
+    # default device
+    tensor_module = ModuleMap(
+        single_block_tensor_torch.keys, modules, out_properties=out_properties
+    )
+    tensor_module.to("cuda")
+    single_block_tensor_torch = single_block_tensor_torch.to(device="cuda")
+    with torch.no_grad():
+        out_tensor = tensor_module(single_block_tensor_torch)
+
+    assert out_tensor.device.type == "cuda"